# api/message_bus/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS Message Bus API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
        <img alt="CasaOS" src="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
    </picture>

    CasaOS Message Bus accepts events and actions from various sources and delivers them to subscribers.

    For issues and discussions, please visit the [GitHub repository](https://github.com/IceWhaleTech/CasaOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/message_bus
...
security:
  - access_token: []
paths:
...
  /ticket:
    post:
      summary: Issue a subscription ticket
      description: |
        Browsers cannot set an `Authorization` header on a WebSocket, and tokens are never accepted
        in URLs (they leak into logs, history and referers). A browser therefore asks for a ticket
        first, with its normal `Authorization` header, and then opens the subscription.

        The ticket is returned only as a cookie: random, valid for one subscription handshake and
        at most 30 seconds, bound to the user and User-Agent, `HttpOnly`, `SameSite=Strict`, and
        scoped to this API's path, so the browser sends it with the WebSocket handshake and nowhere
        else. The same pattern as the root service's SSH terminal login.
      operationId: issueSubscriptionTicket
      tags:
        - SocketIO Endpoints
        - Websocket Endpoints
      responses:
        "204":
          description: Ticket issued in the `Set-Cookie` header.
        "403":
          $ref: "#/components/responses/ResponseForbidden"
        "429":
          $ref: "#/components/responses/ResponseTooManyRequests"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
...
  responses:
...
    ResponseForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Forbidden"
    ResponseTooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Too Many Requests"
...
  schemas:
    BaseResponse:
      properties:
        message:
          readOnly: true
          description: message returned by server side if there is any
          type: string
          example: ""
...